Remove template overrides, move getArguments
