+++ src/index.ts
@@ -1,3 +1,5 @@
+import yargs from 'yargs-parser'
+
 import { RunnerResult, RunnerConfig, RunnerArgs } from './types'
 import resolve from './config-resolver'
 import Logger from './logger'
@@ -5,14 +7,41 @@ import engine, { ShowHelpError } from './engine'
 
 import { printHelp, availableActions, VERSION } from './help'
 
+const getRunnerArgs = (argv: RunnerArgs | string[]): RunnerArgs => {
+  if (Array.isArray(argv)) {
+    const parsed = yargs(argv)
+    const [generator, action, name] = parsed._
+    const { _, ...args } = parsed
+    const [mainAction, subaction] = action.split(':')
+
+    return {
+      generator,
+      action,
+      mainAction,
+      subaction,
+      name,
+      args,
+    }
+  }
+
+  const [mainAction, subaction] = argv.action.split(':')
+
+  return {
+    mainAction,
+    subaction,
+    ...argv,
+  }
+}
+
 const runner = async (
-  argv: RunnerArgs|string[],
+  argv: RunnerArgs | string[],
   config: RunnerConfig,
 ): Promise<RunnerResult> => {
   const resolvedConfig = await resolve(config)
   const { templates, logger } = resolvedConfig
   try {
-    const actions = await engine(argv, resolvedConfig)
+    const runnerArgs = getRunnerArgs(argv)
+    const actions = await engine(runnerArgs, resolvedConfig)
     return { success: true, actions, time: 0 }
   } catch (err) {
     logger.log(err.toString())
@@ -28,4 +57,13 @@ const runner = async (
   }
 }
 
-export { runner, engine, resolve, printHelp, availableActions, Logger, VERSION }
+export {
+  runner,
+  engine,
+  resolve,
+  printHelp,
+  getRunnerArgs,
+  availableActions,
+  Logger,
+  VERSION,
+}
